+++ src/ska_tmc_centralnode/manager/component_manager.py
@@ -283,6 +283,7 @@ class CNComponentManager(TmcComponentManager):
                     [
                         "dishMode",
                         "kValueValidationResult",
+                        "healthState",
                         "longrunningcommandresult",
                         "gpmVersion",
                     ]